# schnorr: Use PutBytesUnchecked for serialize.