--- a/dcrec/secp256k1/schnorr/signature_bench_test.go
+++ b/dcrec/secp256k1/schnorr/signature_bench_test.go
@@ -90,3 +90,23 @@ func BenchmarkSigVerify(b *testing.B) {
 		sig.Verify(msgHash, pubKey)
 	}
 }
+
+// BenchmarkSigVerify benchmarks how long it takes to serialize Schnorr
+// signatures.
+func BenchmarkSigSerialize(b *testing.B) {
+	// From randomly generated keypair.
+	d := hexToModNScalar("9e0699c91ca1e3b7e3c9ba71eb71c89890872be97576010fe593fbf3fd57e66d")
+	privKey := secp256k1.NewPrivateKey(d)
+
+	// blake256 of []byte{0x01, 0x02, 0x03, 0x04}.
+	msgHash := hexToBytes("c301ba9de5d6053caad9f5eb46523f007702add2c62fa39de03146a36b8026b7")
+
+	// Generate the signature.
+	sig, _ := Sign(privKey, msgHash)
+
+	b.ReportAllocs()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		sig.Serialize()
+	}
+}