**This requires #2154**.

schnorr: Use PutBytesUnchecked for serialize.

This updates the code to use the recently-added `PutBytesUnchecked` of the `FieldVal` and `ModNScalar` types to serialize directly into larger buffers where possible to avoid the need for additional copies.

The following benchmarks show a before and after comparison of the affected operations:

```
benchmark               old ns/op    new ns/op    delta
--------------------------------------------------------
BenchmarkSign           52682        52284        -0.76%
BenchmarkSigVerify      178050       176977       -0.60%
BenchmarkSigSerialize   87.5         73.3         -16.23%

benchmark               old allocs   new allocs   delta
--------------------------------------------------------
BenchmarkSign           16           16           +0.00%
BenchmarkSigVerify      21           21           +0.00%
BenchmarkSigSerialize   1            1            +0.00%

benchmark               old bytes    new bytes    delta
--------------------------------------------------------
BenchmarkSign           960          960          +0.00%
BenchmarkSigVerify      1105         1105         +0.00%
```